# Conflicts:
#	Biovation.Brands.EOS/Controllers/EOSDeviceController.cs
#	Biovation.Brands.Suprema/Controllers/SupremaDeviceController.cs
#	Biovation.Brands.ZK/Controllers/ZKAccessGroupController.cs
#	Biovation.Brands.ZK/Controllers/ZKDeviceController.cs
#	Biovation.Brands.ZK/Controllers/ZKLogController.cs
#	Biovation.Brands.ZK/Controllers/ZKTimeZoneController.cs
#	Biovation.Brands.ZK/Controllers/ZKUserController.cs
#	Biovation.Brands.Zk/Controllers/ZKAccessGroupController.cs
#	Biovation.Brands.Zk/Controllers/ZKDeviceController.cs
#	Biovation.Brands.Zk/Controllers/ZKLogController.cs
#	Biovation.Brands.Zk/Controllers/ZKTimeZoneController.cs
#	Biovation.Brands.Zk/Controllers/ZKUserController.cs

+++ Biovation.Data.Commands/Controllers/v2/UserController.cs
@@ -46,7 +46,7 @@ namespace Biovation.Data.Commands.Controllers.v2
         [HttpPost]
         [Route("/DeleteUsers")]
         [Authorize]
-        public Task<ResultViewModel> DeleteUsers([FromBody]List<int> ids = default)
+        public Task<ResultViewModel> DeleteUsers([FromBody]List<long> ids = default)
         {
             return Task.Run(() => _userRepository.DeleteUsers(ids));
         }
